% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catherines_functions.R
\name{pct_chg}
\alias{pct_chg}
\title{Percent Change Calculation}
\usage{
pct_chg(num1, num2)
}
\arguments{
\item{num1}{first number}

\item{num2}{second number}
}
\value{
The percent of change over time
}
\description{
Percent Change Calculation
}
