% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catherines_functions.R
\name{pct_diff}
\alias{pct_diff}
\title{Percent Difference Calculation}
\usage{
pct_diff(num1, num2)
}
\arguments{
\item{num1}{one comparison number}

\item{num2}{the other comparison number}
}
\value{
The percent of difference between two numbers
}
\description{
Percent Difference Calculation
}
